pyproject.toml
[project]
name = "eventyay-unified"
version = "0.1.0"
description = "Platform to manage events (ticketing, talk schedule)."
readme = "README.rst"
requires-python = ">=3.12, <3.13"
keywords = ["tickets", "events", "web", "shop", "ecommerce"]
authors = [ {name = "eventyay", email = "info@eventyay.com"}, ]
maintainers = [ {name = "eventyay", email = "info@eventyay.com"}, ]
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Intended Audience :: Other Audience",
  "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
  "Environment :: Web Environment",
  "License :: OSI Approved :: Apache License, Version 2.0",
  "Programming Language :: Python :: 3.12",
  "Framework :: Django :: 5.2",
]
dependencies = [
    "aiohappyeyeballs>=2.4.4",
    "aiohttp>=3.13.3",
    "aiosignal>=1.3.2",
    "amqp>=5.3.1",
    "annotated-types>=0.7.0",
    "arabic-reshaper>=3.0.0",
    "argon2-cffi>=23.1.0",
    "argon2-cffi-bindings>=25.1.0",
    "asgiref>=3.11.0",
    "asn1crypto>=1.5.1",
    "attrs>=25.4.0",
    "autobahn>=25.12.2",
    "automat>=25.4.16",
    "babel>=2.16.0",
    "beautifulsoup4>=4.14.3",
    "billiard>=4.2.4",
    "bleach>=6.3.0",
    "cbor2>=5.8.0",
    "celery>=5.4.0",
    "certifi>=2026.1.4",
    "cffi>=2.0.0",
    "channels>=4.3.2",
    "channels-redis>=4.2.1",
    "chardet>=5.2.0",
    "charset-normalizer>=3.4.4",
    "click>=8.3.1",
    "click-didyoumean>=0.3.1",
    "click-plugins>=1.1.1",
    "click-repl>=0.3.0",
    "constantly>=23.10.4",
    "contourpy>=1.3.2",
    "cron-descriptor>=2.0.6",
    "cryptography>=44.0.3",
    "css-inline>=0.19.0",
    "csscompressor>=0.9.5",
    "cssutils>=2.11.1",
    "cycler>=0.12.1",
    "daphne>=4.2.1",
    "defusedcsv>=3.0.0",
    "defusedxml>=0.7.1",
    "dj-static>=0.0.6",
    "django>=5.2.5",
    "django-allauth>=65.13.1",
    "django-appconf>=1.2.0",
    "django-bootstrap3>=26.1",
    "django-celery-beat>=2.7.0",
    "django-compressor>=4.6.0",
    "django-context-decorator>=1.6.1",
    "django-cors-headers>=4.9.0",
    "django-countries>=8.2.0",
    # Needed by DRF extensions
    "django-crispy-forms>=2.5",
    "django-csp>=3.8",
    "django-filter>=25.2",
    "django-formset-js-improved>=0.5.0.4",
    "django-formtools>=2.5.1",
    "django-hierarkey>=2.0.1",
    "django-hijack>=3.7.6",
    "django-i18nfield>=1.11.0",
    "django-jquery-js>=3.1.1",
    "django-libsass>=0.9",
    "django-localflavor>=4.0",
    "django-markup>=1.9.1",
    "django-multifactor>=0.8.2",
    "django-oauth-toolkit>=3.2.0",
    "django-otp>=1.7.0",
    "django-pdb>=0.6.2",
    "django-phonenumber-field>=8.4.0",
    "django-redis>=5.4.0",
    "django-scopes>=2.0.0",
    "django-sso>=3.0.2",
    "django-statici18n>=2.5.0",
    "django-timezone-field>=7.2.1",
    "djangorestframework>=3.15.2",
    "drf-flex-fields>=1.0.0",
    "drf-spectacular>=0.28.0",
    "elementpath>=5.1.0",
    "emoji>=2.15.0",
    "et-xmlfile>=2.0.0",
    "eventyay-paypal @ git+https://github.com/fossasia/eventyay-paypal.git@main",
    "eventyay-stripe @ git+https://github.com/fossasia/eventyay-stripe.git@main",
    "eventyay-bitpay @ git+https://github.com/fossasia/eventyay-bitpay.git@main",
    "exhibition @ git+https://github.com/fossasia/eventyay-exhibition.git@main",
    "execnet>=2.1.2",
    "fido2>=2.1.0",
    "fonttools>=4.61.1",
    "frozenlist>=1.8.0",
    "geoip2>=4.8.1",
    "gunicorn>=23.0.0",
    "hyperlink>=21.0.0",
    "icalendar>=6.3.1",
    "idna>=3.11",
    "importlib-metadata>=8.7.1",
    "incremental>=24.11.0",
    "isoweek>=1.3.3",
    "jinja2>=3.1.6",
    "jsonschema>=4.26.0",
    "jsonschema-specifications>=2025.9.1",
    "jwcrypto>=1.5.6",
    "kiwisolver>=1.4.8",
    "kombu>=5.6.2",
    "libsass>=0.23.0",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "markdownify>=1.2.2",
    "matplotlib>=3.10.8",
    "maxminddb>=3.0.0",
    "more-itertools>=10.8.0",
    "msgspec>=0.20.0",
    "mt-940>=4.30.0",
    "multidict>=6.7.0",
    "nh3>=0.3.2",
    "oauthlib>=3.2.2",
    "openpyxl>=3.1.5",
    "orjson>=3.10.18",
    "packaging>=25.0",
    "pandas>=2.3.3",
    "pdf2image>=1.17.0",
    "pdfrw>=0.4",
    "phonenumberslite>=9.0.21",
    "pillow>=12.1.0",
    "polib>=1.2.0",
    "potypo>=0.0.2",
    "protobuf>=6.33.4",
    "psycopg[binary,pool]>=3.3.2",
    "publicsuffixlist>=1.0.2.20260114",
    "pyasn1>=0.6.1",
    "pyasn1-modules>=0.4.2",
    "pycodestyle>=2.12.1",
    "pycountry>=24.6.1",
    "pycparser>=2.23",
    "pycryptodome>=3.20.0",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pydantic-settings>=2.12.0",
    "pygments>=2.19.1",
    "pyjwt>=2.8.0",
    "pyopenssl>=25.3.0",
    "pyotp>=2.9.0",
    "pypdf>=5.1.0",
    "pypng>=0.20220715.0",
    "python-bidi>=0.6.7",
    "python-creole>=1.4.10",
    "python-crontab>=3.2.0",
    "python-dateutil>=2.9.0.post0",
    "python-stdnum>=2.2",
    "pytz>=2025.2",
    "pyuca>=1.2",
    "pyvat>=1.3.18",
    "pyyaml>=6.0.3",
    "qrcode>=8.2",
    "rcssmin>=1.1.2",
    "redis>=7.1.0",
    "regex>=2026.1.15",
    "reportlab>=4.4.9",
    "requests>=2.32.5",
    "requests-oauthlib>=2.0.0",
    "responses>=0.25.5",
    "rich>=14.2.0",
    "rjsmin>=1.2.2",
    "rpds-py>=0.22.3",
    "rules>=3.5",
    "sendgrid>=6.12.5",
    "sentry-sdk>=2.49.0",
    "sepaxml>=2.7.0",
    "service-identity>=24.2.0",
    "sortedcontainers>=2.4.0",
    "soupsieve>=2.8.1",
    "sqlparse>=0.5.5",
    "starkbank-ecdsa>=2.2.0",
    # Hold back by eventyay-stripe
    "stripe>=11.3.0",
    "text-unidecode>=1.3",
    "textile>=4.0.3",
    "tlds>=2025102200",
    "tqdm>=4.67.1",
    "twisted>=25.5.0",
    "txaio>=25.12.2",
    "typing-extensions>=4.15.0",
    "tzdata>=2025.3",
    "urllib3>=2.6.3",
    "urlman>=2.0.3",
    # This library prevents us from being compatible with Python v3.13+.
    "vat-moss-forked>=2020.3.20.0.11.0",
    "vine>=5.1.0",
    "vobject>=0.9.9",
    "wcwidth>=0.2.13",
    "webauthn>=2.2.0",
    "webencodings>=0.5.1",
    "websockets>=15.0.1",
    "xmlschema>=3.4.3",
    "yarl>=1.22.0",
    "zipp>=3.21.0",
    "zxcvbn>=4.4.28",
]

[build-system]
requires = [
  'setuptools>=80.10.2',
  'setuptools-rust',
]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "django-extensions>=4.1",
    "wlc>=1.17.1",
]
docs = [
    "docutils>=0.22.4",
    "pyenchant>=3.3.0",
    "pygments-markdown-lexer>=0.1.0.dev39",
    "sphinx>=9.0.4",
    "sphinx-rtd-theme>=3.1.0",
    "sphinxcontrib-httpdomain>=1.8.1",
    "sphinxcontrib-images>=1.0.1",
    "sphinxcontrib-spelling>=8.0.2",
    "sphinxemoji>=0.3.2",
]
lint = [
    "ruff>=0.14.13",
]
test = [
    "coverage>=7.13.1",
    "coveralls>=4.0.2",
    "faker>=40.1.2",
    "fakeredis>=2.33.0",
    "freezegun>=1.5.5",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cache>=1.0",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.9.0",
    "pytest-env>=1.2.0",
    "pytest-mock>=3.15.1",
    "pytest-rerunfailures>=16.1",
    "pytest-sugar>=1.1.1",
    "pytest-xdist>=3.6.1",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "PLE", "PLW", "UP"]
ignore = ["E722", "E741"]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
# These files re-export models.
"**/__init__.py" = ["F401"]
# These files have long lines.
"**/migrations/*.py" = ["E501"]

[tool.ruff.lint.isort]
lines-after-imports = 2

[tool.setuptools.packages.find]
where = ["src"]
include = ["eventyay*"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest]
DJANGO_SETTINGS_MODULE = "eventyay.config.settings"
testpaths = ["tests"]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::PendingDeprecationWarning",
]

[tool.pytest_env]
EVY_RUNNING_ENVIRONMENT = "testing"

[tool.pyrefly]
project-includes = [
    "**/*.py*",
    "**/*.ipynb",
]
